vms: installing kibana-4.0.2-linux-x64.tar.gz on Debian Jessie 8.0

+++ scripts/elastic.sh
@@ -15,29 +15,28 @@ if [[ ! -n "`grep  ${APT_URL} /etc/apt/sources.list`" ]]; then
     apt-get -y update
 fi
 
-apt-get -y install openjdk-7-jre-headless curl
-apt-get -y install elasticsearch
+if [[ ! -e  /etc/elasticsearch/ ]]; then
+    apt-get -y install openjdk-7-jre-headless curl
+    apt-get -y install elasticsearch
 
-# elasticsearch-1.5.1.deb does not create all directories
-if [[ ! -e /usr/share/elasticsearch/config ]]; then
-    mkdir -p /usr/share/elasticsearch/config
-    ln -s /etc/elasticsearch/elasticsearch.yml /usr/share/elasticsearch/config/elasticsearch.yml
-    ln -s /etc/elasticsearch/logging.yml /usr/share/elasticsearch/config/logging.yml
+    # elasticsearch-1.5.1.deb does not create all directories
+    if [[ ! -e /usr/share/elasticsearch/config ]]; then
+        mkdir -p /usr/share/elasticsearch/config
+        ln -s /etc/elasticsearch/elasticsearch.yml /usr/share/elasticsearch/config/elasticsearch.yml
+        ln -s /etc/elasticsearch/logging.yml /usr/share/elasticsearch/config/logging.yml
+    fi
+
+    if [[ ! -e /usr/share/elasticsearch/data ]]; then
+        mkdir -p /usr/share/elasticsearch/data
+        chown elasticsearch:elasticsearch /usr/share/elasticsearch/data
+    fi
+
+    if [[ ! -e /usr/share/elasticsearch/logs ]]; then
+        ln -s /var/log/elasticsearch /usr/share/elasticsearch/logs
+    fi
+
+    systemctl daemon-reload
+    systemctl enable elasticsearch.service
+    systemctl start elasticsearch.service
 fi
-
-if [[ ! -e /usr/share/elasticsearch/data ]]; then
-    mkdir -p /usr/share/elasticsearch/data
-    chown elasticsearch:elasticsearch /usr/share/elasticsearch/data
-fi
-
-if [[ ! -e /usr/share/elasticsearch/logs ]]; then
-    ln -s /var/log/elasticsearch /usr/share/elasticsearch/logs
-fi
-
-systemctl daemon-reload
-systemctl enable elasticsearch.service
-systemctl start elasticsearch.service
-
-
-
